Moving AgentInstallSelector from salus-telemetry-model
