+++ src/main/java/com/rackspace/salus/telemetry/etcd/lombok.config
@@ -1,2 +1,3 @@
 config.stopBubbling = true
+lombok.accessors.chain = true
 lombok.addLombokGeneratedAnnotation = true
